Address review feedback

+++ internal/metrics/source/source.go
@@ -51,14 +51,14 @@ type VolumeInfo struct {
 var ErrCapacityIsZero = errors.New("capacity is zero")
 
 // FreeSpacePercentage returns the free space as a percentage.
-func (vi *VolumeInfo) FreeSpacePercentage() (float64, error) {
+func (vi *VolumeInfo) FreeSpacePercentage() (int, error) {
 	if vi.CapacityBytes == 0 {
-		return 0.0, ErrCapacityIsZero
+		return 0, ErrCapacityIsZero
 	}
 
 	val := float64(vi.AvailableBytes) / float64(vi.CapacityBytes) * 100.0
 
-	return val, nil
+	return int(val), nil
 }
 
 // UsedSpacePercentage returns the used space as a percentage.
@@ -67,19 +67,19 @@ func (vi *VolumeInfo) UsedSpacePercentage() (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	val := 100.0 - free
+	val := 100 - free
 
-	return int(val), nil
+	return val, nil
 }
 
 // FreeInodesPercentage returns the number of free inodes as a percentage.
-func (vi *VolumeInfo) FreeInodesPercentage() (float64, error) {
+func (vi *VolumeInfo) FreeInodesPercentage() (int, error) {
 	if vi.CapacityInodes == 0 {
-		return 0.0, ErrCapacityIsZero
+		return 0, ErrCapacityIsZero
 	}
 	val := float64(vi.AvailableInodes) / float64(vi.CapacityInodes) * 100.0
 
-	return val, nil
+	return int(val), nil
 }
 
 // UsedInodesPercentage returns the number of used inodes as a percentage.
@@ -88,9 +88,9 @@ func (vi *VolumeInfo) UsedInodesPercentage() (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	val := 100.0 - free
+	val := 100 - free
 
-	return int(val), nil
+	return val, nil
 }
 
 // Metrics is a collection of metrics about persistent volume claims grouped by
